requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_chroma
langchain-openai
langchain-community
langchain-core
gunicorn
pydantic
beautifulsoup4